+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### bug fixes
 - [Ruby] Fix of missing initializer at sprockets. ([#371](https://github.com/fnando/i18n-js/pull/371))
+- [JS] Correctly round unprecise floating point numbers.
 
 
 ## 3.0.0.rc11